refactor: reorganize Tooltip component and types
