+++ src/vis-components/phrases/ui/Tooltip.tsx
@@ -1,8 +1,41 @@
-import { FunctionalComponent, cloneElement, isValidElement, render } from 'preact';
+import { ComponentChildren, FunctionalComponent, cloneElement, isValidElement, render } from 'preact';
 import { useEffect, useRef, useState } from 'preact/hooks';
 import { JSX } from 'preact/jsx-runtime';
 import { getPrefixCls } from '../../../utils';
-import { TooltipProps } from '../../types';
+
+// Define tooltip placement.
+export type TooltipPlacement = 'top' | 'right' | 'bottom' | 'left';
+
+// Tooltip props.
+export interface TooltipProps {
+  /** Tooltip content */
+  title: HTMLElement | string | number;
+  /** Whether tooltip is visible */
+  visible?: boolean;
+  /** Default visibility */
+  defaultVisible?: boolean;
+  /** Child elements */
+  children: ComponentChildren;
+  /** Placement */
+  placement?: TooltipPlacement;
+  /** Trigger method */
+  trigger?: 'hover' | 'click';
+  /** Custom style on tooltip div*/
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  style?: Record<string, any>;
+  /** className */
+  className?: string;
+  /** Show arrow */
+  showArrow?: boolean;
+  /** Callback when visibility changes */
+  onVisibleChange?: (visible: boolean) => void;
+  /** Offset distance */
+  offset?: number;
+  /** Mouse enter delay (ms) */
+  mouseEnterDelay?: number;
+  /** Mouse leave delay (ms) */
+  mouseLeaveDelay?: number;
+}
 
 const TOOLTIP_CONTAINER_ID = 'ntv-tooltip-container';
