+++ src/vis-components/types.ts
@@ -1,4 +1,3 @@
-import { ComponentChildren } from 'preact';
 import { NarrativeTextSpec, SectionSpec, ParagraphSpec, PhraseSpec, HeadlineSpec, BulletItemSpec } from '../schema';
 
 export type PhraseEvents = Partial<{
@@ -30,37 +29,3 @@ export type NarrativeEvents = SectionEvents &
     onCopySuccess: (e?: ClipboardEvent) => void;
     onCopyFailure: (e?: ClipboardEvent) => void;
   }>;
-
-// Define tooltip placement.
-export type TooltipPlacement = 'top' | 'right' | 'bottom' | 'left';
-
-// Tooltip props.
-export interface TooltipProps {
-  /** Tooltip content */
-  title: HTMLElement | string | number;
-  /** Whether tooltip is visible */
-  visible?: boolean;
-  /** Default visibility */
-  defaultVisible?: boolean;
-  /** Child elements */
-  children: ComponentChildren;
-  /** Placement */
-  placement?: TooltipPlacement;
-  /** Trigger method */
-  trigger?: 'hover' | 'click';
-  /** Custom style on tooltip div*/
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  style?: Record<string, any>;
-  /** className */
-  className?: string;
-  /** Show arrow */
-  showArrow?: boolean;
-  /** Callback when visibility changes */
-  onVisibleChange?: (visible: boolean) => void;
-  /** Offset distance */
-  offset?: number;
-  /** Mouse enter delay (ms) */
-  mouseEnterDelay?: number;
-  /** Mouse leave delay (ms) */
-  mouseLeaveDelay?: number;
-}
